Go back to using a std::thread for tiles
It appears that the std::future thing I was using doesn't actually
appear to work -- only one of them ever gets executed at a time.  As I
don't fully understand what they're supposed to be for, I instead have
just decided to go ahead and use a std::thread and handle my own
return values.

This is a little bit nasty because I have to handle the return value
with a special variable, but I guess that's good enough for now.